+++ af_ipf/__manifest__.py
@@ -20,7 +20,6 @@ Asök booking management \n
     "data": [
         "views/ipf_views.xml",
         "data/customer_data.xml",
-        "data/ais_a_data.xml",
     ],
     "application": True,
     "installable": True,
