+++ af_ipf/data/ais_a_data.xml
@@ -5,14 +5,13 @@
 		<record id="ipf_ais_a" model="af.ipf">
 			<field name="name">AIS-Å</field>
 			<field name="environment">T2</field>
-			<field name="systemid">AFCRM</field>
+			<field name="systemid">CRM</field>
 			<field name="url">https://ipfapi.arbetsformedlingen.se</field>
 			<field name="port">443</field>
-			<field name="enduserid">True</field>
 		</record>
 		<record id="ipf_endpoint_ais_a" model="af.ipf.endpoint">
-			<field name="name">ais-beslut-om-stod-read/v1/arenden/sokande/{personnummer}?client_id={clientid}&amp;client_secret={client_secret}</field>
-			<field name="ipf_id" ref="ipf_ais_a"/>
+			<field name="name">ais-beslut-om-stod-read/v1/arenden/sokande/{personnummer}?client_id={client_id}&client_secret={client_secret}</field>
+			<field name="ipf_id" ref="ipf_c"/>
 	</record>
 	</data>
 </odoo>
