+++ af_ipf/views/ipf_views.xml
@@ -18,18 +18,17 @@
                 <sheet>
                     <group>
                         <field name="name"/>
-                        <field name="url"/>
-                        <field name="port"/>
                         <field name="clientid"/>
-                        <field name="client_secret" password="True"/>
+                        <field name="client_secret"/>
                         <field name="auth_user"/>
-                        <field name="auth_password" password="True"/>
+                        <field name="auth_password"/>
                         <field name="environment"/>
                         <field name="systemid"/>
-                        <field name="enduserid"/>
+                        <field name="url"/>
+                        <field name="port"/>
                         <field name="ssl_verify"/>
                         <field name="ssl_cert"/>
-                        <field name="ssl_key" password="True"/>
+                        <field name="ssl_key"/>
                     </group>
                     <label for="endpoint_ids"/>
                     <field name="endpoint_ids"/>
