--- ipf_ais_a _2/i18n/sv_SE.po
@@ -1,115 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* ipf_ais_a
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 12.0-20200508\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2020-09-28 06:36+0000\n"
-"PO-Revision-Date: 2020-09-28 16:18+0200\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: \n"
-"Language: sv_SE\n"
-"X-Generator: Poedit 2.0.6\n"
-
-#. module: ipf_ais_a
-#: model_terms:ir.ui.view,arch_db:ipf_ais_a.view_ais_a_tree
-#: model_terms:ir.ui.view,arch_db:ipf_ais_a.view_jobseeker_form
-msgid "AIS-Å"
-msgstr "AIS-Å"
-
-#. module: ipf_ais_a
-#: model:ir.model,name:ipf_ais_a.model_res_partner_ais_a
-msgid "AIS-Å Beslut Om Stod"
-msgstr "AIS-Å Beslut om Stöd"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__dec_office
-msgid "Beslutande Kontor"
-msgstr "Beslutande kontor"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__name
-msgid "Case nr"
-msgstr "Beslutsnummer"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner__ais_a_ids
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_users__ais_a_ids
-msgid "Cases"
-msgstr "Beslut"
-
-#. module: ipf_ais_a
-#: model:ir.model,name:ipf_ais_a.model_res_partner
-msgid "Contact"
-msgstr "Kontakt"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__create_uid
-msgid "Created by"
-msgstr "Skapad av"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__create_date
-msgid "Created on"
-msgstr "Skapad den"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__display_name
-msgid "Display Name"
-msgstr "Visningsnamn"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__res_officer
-msgid "Handlaggare"
-msgstr "Handläggare"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__id
-msgid "ID"
-msgstr "ID"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__partner_id
-msgid "Job seeker"
-msgstr "Arbetssökande"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__res_office
-msgid "Kontor"
-msgstr "Kontor"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a____last_update
-msgid "Last Modified on"
-msgstr "Senast redigerad"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__write_uid
-msgid "Last Updated by"
-msgstr "Senast uppdaterad av"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__write_date
-msgid "Last Updated on"
-msgstr "Senast uppdaterad"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__status_change
-msgid "Senaste Status Ändring"
-msgstr "Senaste statusändring"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__status
-msgid "Status"
-msgstr "Status"
-
-#. module: ipf_ais_a
-#: model:ir.model.fields,field_description:ipf_ais_a.field_res_partner_ais_a__atgard
-msgid "Åtgärd"
-msgstr "Åtgärd"
