--- ipf_ais_a _2/static/src/xml/ais_a.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates>
-
-    <!--
-        Ais-Å tab
-    -->
-
-    <div t-name="ipf_ais_a.ais_a" class="af_ais_a">
-        <div class="af_ais_a">
-            hej hopp!
-        </div>
-    </div>
-</templates>
